Move XComOperatorLink to serialized_objects as only API server uses it
